added Courses and CandidateCourses

+++ src/main/java/com/app/controller/AdminController.java
@@ -24,7 +24,7 @@ import com.app.model.Admin;
 import com.app.service.AdminService;
 
 
-// new admin comment
+
 @RestController
 public class AdminController {
 
@@ -37,7 +37,7 @@ public class AdminController {
 	    }
 
 	    @GetMapping("/admins/{id}")
-	    public ResponseEntity<Optional<Admin>> getAdminById(@PathVariable(value = "id") int adminId)
+	    public ResponseEntity<Optional<Admin>> getAdminById(@PathVariable(value = "id") Long adminId)
 	        throws ResourceNotFoundException {
 	    	Optional<Admin> admin = adminService.findById(adminId);
 	        return ResponseEntity.ok().body(admin);
@@ -49,7 +49,7 @@ public class AdminController {
 	    }
 
 	    @PutMapping("/admins/{id}")
-	    public ResponseEntity<Admin> updateAdmin(@PathVariable(value = "id") int adminsId,@Valid @RequestBody Admin adminDetails) throws ResourceNotFoundException {
+	    public ResponseEntity<Admin> updateAdmin(@PathVariable(value = "id") Long adminsId,@Valid @RequestBody Admin adminDetails) throws ResourceNotFoundException {
 	    	Admin admin = adminService.findById(adminsId)
 	        .orElseThrow(() -> new ResourceNotFoundException("admin not found for this id :: " + adminsId));
 
@@ -63,7 +63,7 @@ public class AdminController {
 	    }
 
 	    @DeleteMapping("/admins/{id}")
-	    public Map<String, Boolean> deleteAdmin(@PathVariable(value = "id") int adminId)
+	    public Map<String, Boolean> deleteAdmin(@PathVariable(value = "id") Long adminId)
 	         throws ResourceNotFoundException {
 	    	Admin admin = adminService.findById(adminId)
 	       .orElseThrow(() -> new ResourceNotFoundException("Admin not found for this id :: " + adminId));
