+++ src/main/java/com/app/repo/HostelRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.app.model.Hostel;
 
-public interface HostelRepository extends JpaRepository<Hostel,Integer> {
+public interface HostelRepository extends JpaRepository<Hostel,Long> {
 	
 }
